add: tests for modal window, refactored files and file names
